API/requirements.txt
Flask
pybase64
Pillow
opencv-python
pytesseract
pylev
numpy